refactor to avoid compilation redundancy
